Update docs: new() will no longer be part of the net adapter API
The new() method is not required for an HTTP adapter because the plug-in factory event handler will be in charge of instantiating the object. It may do so using new(), but it may also choose another option. However, for the soon-to-be deprecated net_module option, new() is still required.

Additionally, the net adapter API docs were moved from Net to Plugin.